Se calculan los ciclos de demora por trafico

+++ headers/Arista.h
@@ -13,7 +13,7 @@ class Arista {
 public:
     int ancho_de_banda;     //Ancho de banda en MegaByte. Especificado en archivo txt entre 25 MB - 100 MB
     int peso;               //Inversa del ancho de banda, para definir los caminos mas cortos inciales
-    int ciclos;             //Ciclos de demora debido al trafico
+    double ciclos;             //Ciclos de demora debido al trafico
     Router* destino;
     Arista* next;
